+++ util/opentelemetry-util-genai/src/opentelemetry/util/genai/types.py
@@ -204,6 +204,23 @@ class GenAIInvocation:
 
 
 @dataclass
+class WorkflowInvocation(GenAIInvocation):
+    """
+    Represents predetermined static sequence of operations eg: Agent, LLM, tool, and retrieval invocations.
+    A workflow groups multiple operations together, accepting input(s) and producing final output(s).
+    """
+
+    name: str = ""
+    operation_name: str = "invoke_workflow"
+    input_messages: list[InputMessage] = field(
+        default_factory=_new_input_messages
+    )
+    output_messages: list[OutputMessage] = field(
+        default_factory=_new_output_messages
+    )
+
+
+@dataclass
 class LLMInvocation(GenAIInvocation):
     """
     Represents a single LLM call invocation. When creating an LLMInvocation object,